backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
langchain
langchain-google-genai
google-generativeai
langchain-community
faiss-cpu
pydantic
